refactor: remove residual dead runtime paths
Delete abandoned pricing, admin, blacklist and direct-digest paths so tests and consumers exercise the live implementation owners. Collapse redundant reexports and relocate test-only utilities without counting relocation as source reduction. Preserve production pricing timestamps, durable Telegram delivery and retained feature configuration; correct stale ownership docs.

Plan: agents/astra-code-deslop/IMPLEMENTATION-PLAN.md
Tasks: W1.1, W1.2, W1.3, W1.4, W1.5, W1.6, W1.7, W1.8, FR-A3, DY-5
Verification: 59 suites / 460 tests; restored Telegram contract 56 tests; exact primary-price parity; root/Worker types; ESLint; source, docs, cron and clone gates.
