--- worker/src/api/status-supplements.ts
@@ -1,1 +1,0 @@
-export { loadStatusSupplements } from "../lib/status/supplements";
