+++ worker/src/cron/sync-stablecoins/enrich-prices-primary-shared.ts
@@ -1,5 +1,3 @@
-import { getPricingSourceRegistryEntry } from "@shared/lib/pricing-source-registry";
-
 export interface PriceValidationStats {
   attempted: number;
   high: number;
@@ -14,8 +12,4 @@ const INVALID_GECKO_ID_SENTINEL = "wrong";
 
 export function isUsableGeckoId(geckoId: unknown): geckoId is string {
   return typeof geckoId === "string" && geckoId.length > 0 && geckoId !== INVALID_GECKO_ID_SENTINEL;
-}
-
-export function getSourceDefaultWeight(source: string): number {
-  return getPricingSourceRegistryEntry(source)?.defaultWeight ?? 1;
 }
