--- worker/src/cron/yield-sync/sources-optional-protocols.ts
@@ -1,31 +1,0 @@
-export {
-  BASEDOLLAR_SP_CONFIG,
-  LIQUITY_V2_SP_CONFIG,
-  fetchBprotocolLqtyOnlySource,
-  fetchCurveScrvusdCurrentRateSource,
-  fetchLiquityV2StabilityPoolSource,
-  type LiquityV2SpSourceConfig,
-} from "./sources-optional-protocols-onchain";
-export {
-  fetchBimaSusbdSource,
-  fetchEtherfuseCetesSource,
-  fetchHashnoteUsycSource,
-  fetchOndoUsdyOracleSource,
-  fetchReProtocolReusdSource,
-  fetchYearnYboldSource,
-} from "./sources-optional-protocols-protocol-api";
-export { fetchZephyrZysSource } from "../../lib/yield-source-adapters/zephyr";
-export {
-  fetchRoycoDawnSources,
-} from "./royco-dawn";
-export {
-  fetchVaultsFyiSources,
-  type VaultsFyiSourceResult,
-  type VaultsFyiTelemetry,
-} from "./vaults-fyi";
-export {
-  fetchBeefySources,
-  fetchMorphoVaultSources,
-  fetchPendleMarketSources,
-  fetchYearnKongSources,
-} from "./sources-optional-protocols-supplemental";
