+++ worker/src/lib/primary-price-collector.ts
@@ -1,7 +1,6 @@
 import type { PriceObservedAtMode, PriceSourceConfidenceProfile } from "@shared/types/core";
 import { getPricingSourceRegistryEntry } from "@shared/lib/pricing-source-registry";
 import { DIVERGENCE_THRESHOLD_BPS } from "@shared/lib/pricing-pipeline-constants";
-import type { AddressPriceQuote } from "./address-price-providers";
 import { pricesAgreeWithinBps } from "./price-divergence";
 import type { SourcePrice } from "./price-consensus";
 import { validatePricingSourceFreshness, type PricingSourceFreshnessRejectReason } from "./pricing-source-freshness";
@@ -115,7 +114,6 @@ export interface PrimaryCollectedQuotes {
   curveOraclePrice: number | null;
   curveOracleObservedAt: number | null;
   navQuote?: NavTelemetryQuote;
-  addressProviderQuotes?: AddressPriceQuote[];
   protocolSources?: DexProtocolSourceQuote[];
   dexAggregateQuote?: DexAggregateQuote;
 }
@@ -308,27 +306,6 @@ export function buildPrimarySourceCandidates(
     });
     if (redstoneSource) {
       sources.push(redstoneSource);
-    }
-  }
-
-  for (const quote of collected.addressProviderQuotes ?? []) {
-    const source = buildSourcePrice({
-      source: quote.source,
-      price: quote.priceUsd,
-      observedAt: quote.observedAt,
-      observedAtMode: quote.observedAtMode,
-      nowSec,
-      metadata: {
-        chain: quote.chain,
-        address: quote.address,
-        liquidityUsd: quote.liquidityUsd,
-        volume24hUsd: quote.volume24hUsd,
-        poolCount: quote.poolCount,
-        ...quote.metadata,
-      },
-    });
-    if (source) {
-      sources.push(source);
     }
   }
