+++ worker/src/lib/telegram.ts
@@ -1,7 +1,6 @@
 import { parseRetryAfterSeconds } from "@shared/lib/retry-after";
 import { TELEGRAM_BOT_URL } from "@shared/lib/telegram-bot-registration";
 import type { TelegramRecapRolloutPolicy } from "@shared/lib/telegram-recap-rollout";
-import { logWorkerEventArgs } from "./structured-log";
 import { drainResponseBody, readResponseTextBoundedWithSignal } from "./response-body";
 import { escapeHtml } from "./telegram/html";
 import { logTelegramEvent } from "./telegram/log";
@@ -126,40 +125,6 @@ export function buildTelegramMessage(
     recapCta ?? "",
   ].filter((section) => section.trim().length > 0);
   return sections.join("\n\n");
-}
-
-/** Post a raw text message to a Telegram channel. Throws on API error. */
-export async function postTelegramMessage(text: string, creds: TelegramCreds): Promise<void> {
-  const result = await sendToChat(creds.chatId, text, creds.botToken);
-  if (!result.ok) {
-    throw new Error(`Telegram API ${result.statusCode ?? "?"}: ${result.errorClass}`);
-  }
-}
-
-/**
- * Format and post a digest to the Telegram channel.
- * The caller is responsible for catching errors (this is non-fatal).
- */
-export async function postDigestToTelegram(
-  title: string,
-  extended: string,
-  date: string,
-  creds: TelegramCreds,
-  editionNumber?: number | null,
-  appendixHtml?: string | null,
-  recapRollout?: TelegramRecapRolloutPolicy | null,
-): Promise<void> {
-  const text = buildTelegramMessage(
-    title,
-    extended,
-    date,
-    editionNumber,
-    appendixHtml,
-    null,
-    recapRollout,
-  );
-  await postTelegramMessage(text, creds);
-  logWorkerEventArgs("lib", "info", `[telegram] Posted digest (${text.length} chars)`);
 }
 
 /**
